* EqualityEpsilon for Equals method is Double, fixed SFNode and MFNode
git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/kambi_vrml_game_engine@2794 afe8dcc5-9094-451b-8773-5f495042a737

+++ 3dmodels/vrmlnodes.pas
@@ -1490,7 +1490,7 @@ type
     procedure Parse(Lexer: TVRMLLexer; IsClauseAllowed: boolean); override;
     function EqualsDefaultValue: boolean; override;
     function Equals(SecondValue: TVRMLField;
-      const EqualityEpsilon: Single): boolean; override;
+      const EqualityEpsilon: Double): boolean; override;
     procedure Assign(Source: TPersistent); override;
     procedure AssignValue(Source: TVRMLField); override;
 
@@ -1590,7 +1590,7 @@ type
 
     function EqualsDefaultValue: boolean; override;
     function Equals(SecondValue: TVRMLField;
-      const EqualityEpsilon: Single): boolean; override;
+      const EqualityEpsilon: Double): boolean; override;
     procedure Assign(Source: TPersistent); override;
     procedure AssignValue(Source: TVRMLField); override;
 
@@ -4386,7 +4386,7 @@ begin
 end;
 
 function TSFNode.Equals(SecondValue: TVRMLField;
-  const EqualityEpsilon: Single): boolean;
+  const EqualityEpsilon: Double): boolean;
 begin
  Result := (inherited Equals(SecondValue, EqualityEpsilon)) and
    (SecondValue is TSFNode) and
@@ -4617,7 +4617,7 @@ begin
 end;
 
 function TMFNode.Equals(SecondValue: TVRMLField;
-  const EqualityEpsilon: Single): boolean;
+  const EqualityEpsilon: Double): boolean;
 begin
   Result := (inherited Equals(SecondValue, EqualityEpsilon)) and
     (SecondValue is TMFNode) and
